requirements.txt
requests>=2.28
beautifulsoup4>=4.12
lxml>=4.9
python-dotenv>=1.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
telethon>=1.35.0
python-dateutil>=2.8.0
